Extract shared coercion and JSON helpers into common types
- Add `keymasq/common/coercion.py` and `keymasq/common/types.py` for reusable coercion/JSON/type utilities
- Replace duplicated local `_int/_float/_str/_bool/_json` helpers across daemon, runtime, session, GUI, CLI, and listeners
- Centralize shared `JsonObject`/`JsonObjectList`/`SyntheticInputEvent` usage to reduce duplication and keep behavior consistent
